+++ spec/support/model_parsers/representable_parser.rb
@@ -279,7 +279,6 @@ RSpec.shared_context 'representable swagger example' do
       'paths' => {
         '/v3/other_thing/{elements}' => {
           'get' => {
-            'summary' => 'nested route inside namespace',
             'description' => 'nested route inside namespace',
             'produces' => ['application/json'],
             'parameters' => [{ 'in' => 'body', 'name' => 'elements', 'description' => 'Set of configuration', 'type' => 'array', 'items' => { 'type' => 'string' }, 'required' => true }],
@@ -292,7 +291,6 @@ RSpec.shared_context 'representable swagger example' do
         },
         '/thing' => {
           'get' => {
-            'summary' => 'This gets Things.',
             'description' => 'This gets Things.',
             'produces' => ['application/json'],
             'parameters' => [
@@ -306,7 +304,6 @@ RSpec.shared_context 'representable swagger example' do
             'operationId' => 'getThing'
           },
           'post' => {
-            'summary' => 'This creates Thing.',
             'description' => 'This creates Thing.',
             'produces' => ['application/json'],
             'consumes' => ['application/json'],
@@ -321,7 +318,6 @@ RSpec.shared_context 'representable swagger example' do
         },
         '/thing/{id}' => {
           'get' => {
-            'summary' => 'This gets Thing.',
             'description' => 'This gets Thing.',
             'produces' => ['application/json'],
             'parameters' => [{ 'in' => 'path', 'name' => 'id', 'type' => 'integer', 'format' => 'int32', 'required' => true }],
@@ -330,7 +326,6 @@ RSpec.shared_context 'representable swagger example' do
             'operationId' => 'getThingId'
           },
           'put' => {
-            'summary' => 'This updates Thing.',
             'description' => 'This updates Thing.',
             'produces' => ['application/json'],
             'consumes' => ['application/json'],
@@ -344,7 +339,6 @@ RSpec.shared_context 'representable swagger example' do
             'operationId' => 'putThingId'
           },
           'delete' => {
-            'summary' => 'This deletes Thing.',
             'description' => 'This deletes Thing.',
             'produces' => ['application/json'],
             'parameters' => [{ 'in' => 'path', 'name' => 'id', 'type' => 'integer', 'format' => 'int32', 'required' => true }],
@@ -355,7 +349,6 @@ RSpec.shared_context 'representable swagger example' do
         },
         '/thing2' => {
           'get' => {
-            'summary' => 'This gets Things.',
             'description' => 'This gets Things.',
             'produces' => ['application/json'],
             'responses' => { '200' => { 'description' => 'get Horses', 'schema' => { '$ref' => '#/definitions/Something' } }, '401' => { 'description' => 'HorsesOutError', 'schema' => { '$ref' => '#/definitions/ApiError' } } },
@@ -365,7 +358,6 @@ RSpec.shared_context 'representable swagger example' do
         },
         '/dummy/{id}' => {
           'delete' => {
-            'summary' => 'dummy route.',
             'description' => 'dummy route.',
             'produces' => ['application/json'],
             'parameters' => [{ 'in' => 'path', 'name' => 'id', 'type' => 'integer', 'format' => 'int32', 'required' => true }],