+++ spec/swagger_v2/mounted_target_class_spec.rb
@@ -39,7 +39,6 @@ describe 'docs mounted separately from api' do
       'paths' => {
         '/simple' => {
           'get' => {
-            'summary' => 'This gets something.',
             'description' => 'This gets something.',
             'produces' => ['application/json'],
             'responses' => { '200' => { 'description' => 'This gets something.' } },
@@ -62,7 +61,6 @@ describe 'docs mounted separately from api' do
       'paths' => {
         '/simple' => {
           'get' => {
-            'summary' => 'This gets something.',
             'description' => 'This gets something.',
             'produces' => ['application/json'],
             'responses' => {